+++ src/Attributes/DataProvider.php
@@ -14,7 +14,12 @@ use Attribute;
 #[Attribute(Attribute::TARGET_METHOD)]
 final readonly class DataProvider
 {
-    public function __construct(public string $value)
+    public function __construct(private string $methodName)
     {
+    }
+
+    public function getValue(): string
+    {
+        return $this->methodName;
     }
 }