+++ src/Attributes/RequiresPhpExtension.php
@@ -15,14 +15,19 @@ use MyTester\ISkipAttribute;
 #[Attribute(Attribute::TARGET_METHOD | Attribute::IS_REPEATABLE)]
 final readonly class RequiresPhpExtension implements ISkipAttribute
 {
-    public function __construct(public string $value)
+    public function __construct(private string $extensionName)
     {
+    }
+
+    public function getValue(): string
+    {
+        return $this->extensionName;
     }
 
     public function getSkipValue(): ?string
     {
-        if (!extension_loaded($this->value)) {
-            return "extension $this->value is not loaded";
+        if (!extension_loaded($this->extensionName)) {
+            return "extension $this->extensionName is not loaded";
         }
         return null;
     }