+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Version 8.0.0-dev
 - BC break: Tester's constructor now takes new class TestSuitesSelectionCriteria as its first parameter
 - deprecated ITestSuitesFinder::FILENAME_SUFFIX
 - BC break: made TestCase::METHOD_PATTERN protected
+- BC break: renamed and made private attributes' properties, added method getValue instead
 
 Version 7.3.1
 - allowed installation konecnyjakub/event-dispatcher 2